-o, '&': Hide files which are on one side only

+++ ed.c
@@ -145,7 +145,7 @@ set_fkey(int i, char *s)
 	free(fkey_cmd[i]);
 	fkey_cmd[i] = NULL;
 
-	if (*s == '$' && isspace(s[1])) {
+	if (*s == '$' && isspace((int)s[1])) {
 		int c;
 		char *p = s;
